+++ vehicle/src/Vehicle/Verify/QueryFormat/VNNLib.hs
@@ -53,7 +53,7 @@ compileAssertion varNames assertion = do
 
   let compiledRel = compileRel rel
   let compiledLHS = foldl compileVar "" (NonEmpty.tail coeffVars)
-  let compiledRHS = prettyCoefficient constant
+  let compiledRHS = prettyRationalAsFloat constant
   return $ parens "assert" <+> parens (compiledRel <+> parens compiledLHS <+> compiledRHS)
 
 compileRel :: Either () OrderOp -> Doc a
@@ -68,5 +68,5 @@ compileVar :: Doc a -> (Coefficient, Name) -> Doc a
 compileVar r (coef, var)
   | coef == 1 = "+" <+> parens r <+> pretty var
   | coef == -1 = "-" <+> parens r <+> pretty var
-  | coef < 0 = "-" <+> parens r <+> parens ("*" <+> prettyCoefficient (-coef) <+> pretty var)
-  | otherwise = "+" <+> parens r <+> parens ("*" <+> prettyCoefficient coef <+> pretty var)
+  | coef < 0 = "-" <+> parens r <+> parens ("*" <+> prettyRationalAsFloat (-coef) <+> pretty var)
+  | otherwise = "+" <+> parens r <+> parens ("*" <+> prettyRationalAsFloat coef <+> pretty var)